# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Karat Hive API",
    "version": "0.1.0",
    "description": "Generated from Nest controllers + ZodValidationPipe schemas (`npm run openapi:generate`). Zod remains the single source of request shape (NFR-030, AD-BE-14). First artefact is partial: paths/methods and request Zod schemas are covered; response presenters are not yet dual-exported as Zod. Pre-code catalogue: docs/API-Route-Inventory.md — clients should treat this file as the evolving contract."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Karat Hive access JWT (not a Firebase ID token)."
      }
    },
    "schemas": {
      "SuccessMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "type": "string"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "nextCursor"
        ]
      },
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMeta"
          }
        },
        "required": [
          "meta"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/reviews": {
      "get": {
        "operationId": "Admin_listReviews",
        "tags": [
          "Admin"
        ],
        "summary": "GET /v1/admin/reviews",
        "description": "Handler Admin.listReviews. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/connections/{id}/reviews": {
      "post": {
        "operationId": "Review_createReview",
        "tags": [
          "Review"
        ],
        "summary": "POST /v1/connections/{id}/reviews",
        "description": "Handler Review.createReview. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 1000
                  }
                },
                "required": [
                  "rating"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/reviews": {
      "get": {
        "operationId": "Review_listMyReviews",
        "tags": [
          "Review"
        ],
        "summary": "GET /v1/me/reviews",
        "description": "Handler Review.listMyReviews. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "AUTHOR",
                "SUBJECT"
              ]
            },
            "required": false,
            "name": "role",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reviews/{id}": {
      "patch": {
        "operationId": "Review_updateReview",
        "tags": [
          "Review"
        ],
        "summary": "PATCH /v1/reviews/{id}",
        "description": "Handler Review.updateReview. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reviews/{id}/flag": {
      "post": {
        "operationId": "Review_flagReview",
        "tags": [
          "Review"
        ],
        "summary": "POST /v1/reviews/{id}/flag",
        "description": "Handler Review.flagReview. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reviews/{id}/response": {
      "post": {
        "operationId": "Review_respondToReview",
        "tags": [
          "Review"
        ],
        "summary": "POST /v1/reviews/{id}/response",
        "description": "Handler Review.respondToReview. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "response": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  }
                },
                "required": [
                  "response"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reviews/{id}/withdraw": {
      "post": {
        "operationId": "Review_withdrawReview",
        "tags": [
          "Review"
        ],
        "summary": "POST /v1/reviews/{id}/withdraw",
        "description": "Handler Review.withdrawReview. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  }
}